Implement options for tab width, tab expansion
Implement an option which controls the number of spaces between tab
stops.  Basically, just turning the tabstop_length constant into a
variable that can be set with a flag value.

Implement support for tab expansion, enabled by a flag.  The
tabstop_length variable is the number of spaces in the "expanded"
tab.  When enabled:
- TAB inserts spaces
- C-x > inserts spaces
- C-x < removes spaces (or a tab or spaces then a tab)
- Backspacing a space will continue to delete spaces until the next
  tab stop.